+++ display/index.js
@@ -2,6 +2,7 @@ const describeConfidence = require('./describe_confidence');
 const describeRoute = require('./describe_route');
 const describeRoutingFailure = require('./describe_routing_failure');
 const formatFeeRate = require('./format_fee_rate');
+const parseAmount = require('./parse_amount');
 const segmentMeasure = require('./segment_measure');
 const sumsForSegment = require('./sums_for_segment');
 
@@ -10,6 +11,7 @@ module.exports = {
   describeRoute,
   describeRoutingFailure,
   formatFeeRate,
+  parseAmount,
   segmentMeasure,
   sumsForSegment,
 };